rust_project::buck

Struct Buck

pub(crate) struct Buck {
    command: String,
    mode: Option<String>,
}

Fields
command: Stringmode: Option<String>
Implementations

impl Buck

pub(crate) fn new(command: Option<String>, mode: Option<String>) -> Self

fn command<I, S>(&self, subcommands: I, xplat_platforms: &[&str]) -> Command
where I: IntoIterator<Item = S>, S: AsRef<OsStr>,

Invoke buck2 with the given subcommands.

Care should be taken to ensure that buck is invoked with the same set options and configuration to avoid invalidating caches.

fn command_without_config<I, S>(&self, subcommands: I) -> Command
where I: IntoIterator<Item = S>, S: AsRef<OsStr>,

Invoke buck with the given subcommands.

This method should only be used with with buck commands that do not accept configuration options, such as root. Buck::command should be preferred.

pub(crate) fn resolve_project_root(&self) -> Result<PathBuf, Error>

Return the absolute path of the current Buck project root.

pub(crate) fn resolve_sysroot_src(&self) -> Result<PathBuf, Error>

pub(crate) fn check_saved_file( &self, use_clippy: bool, saved_file: &Path, ) -> Result<CheckOutput, Error>

Determines the owning target(s) of the saved file and builds them.

pub(crate) fn expand_and_resolve( &self, targets: &[Target], exclude_workspaces: bool, ) -> Result<ExpandedAndResolved>

pub(crate) fn query_aliased_libraries( &self, targets: &[Target], ) -> Result<FxHashMap<Target, AliasedTargetInfo>, Error>

pub(crate) fn query_owners( &self, input: Input, max_extra_targets: usize, ) -> Result<FxHashMap<PathBuf, Vec<Target>>, Error>

Find the buildfile that owns each file specified, and return the path to each buildfile along with all the targets it contains.